Create json_2_csv.py
This script assumes that the input JSON file contains an array of objects, each of which has the same keys. The script extracts the field names from the first object in the array and uses them as the column headers in the output CSV file. Each subsequent object is written to a new row in the CSV file.

Note that the newline='' argument is included in the open() function call when opening the output file to avoid extra blank lines in the resulting CSV file.